+++ src/main/java/com/sbproject/piclocket/dto/PhotoResponse.java
@@ -6,6 +6,7 @@ import java.util.UUID;
 
 public record PhotoResponse(
         UUID photoId,
-        PhotoStatus status
+        PhotoStatus status,
+        String downloadUrl
 ) {
 }